reimplement initializing a MySQL database user
Previously dbsake would inject a user during the mysqld --bootstrap
phase based on either parsing a .frm or munging a line from one
of the MySQL bootstrap scripts.  Now this is more elegantly
handled by just starting the new sandbox with --init-file and
issuing a standard grant statement.

This resolves an issue with MySQL 5.7 (which has changed the
initialization scripts) and makes the sandbox initialization
more robust.