fix datasource progressbar to work with new compression api
compression now loads a separate background thread to track
the progress of loading a datasource and pumps data from
the underlying source into the compression command's stdin.
This is done to count bytes correctly and provide proper
progress.

As part of this, the progressbar itself has been updated to
provide more information including rate and ETA information
similar to "pv".

Resolves issue #73