fix chown to fail more gracefully with python26
with python2.6, chown with a bad type results in a DeprecationWarning
rather than failing with a TypeError. python2.7+ fails fast with
a TypeError, as expected.

This puts a little shim to check the type and fail cleanly if
either user or group results in a non-integer uid before calling
os.chown.