Delay bootstrapping until sandbox.sh has been generated
This pushes the bootstrapping process until after sandbox.sh is
rendered.  This allows recovering from a failed bootstrap process
more easily. This might happen when testing against a newer version
of MySQL that the sandbox command doesn't support yet, or if the
particular datasource used is incompatible with the distribution of
MySQL chosen, among other possible issues.